+++ src/utils/config.ts
@@ -1,10 +1,5 @@
-import dotenv from 'dotenv';
 import { FunctionDefinitionFormat } from './enum.js';
-
-// Only load .env in non-test environments
-if (process.env.NODE_ENV !== 'test') {
-  dotenv.config();
-}
+import { validateApiKey } from './validators.js';
 
 export interface AciMCPConfig {
   readonly ACI_API_KEY: string;
@@ -14,10 +9,9 @@ export interface AciMCPConfig {
   readonly FUNCTION_DEFINITION_FORMAT: string;
 }
 
-export const createAciMCPConfig = (options?: {
-  allowedAppsOnly?: boolean;
-  linkedAccountOwnerId?: string;
-}): AciMCPConfig => {
+let config: AciMCPConfig;
+
+export const initConfig = (options?: { allowedAppsOnly?: boolean; linkedAccountOwnerId?: string }): void => {
   const ACI_SERVER_URL = process.env.ACI_SERVER_URL || 'https://api.aci.dev/v1/';
   const FUNCTION_DEFINITION_FORMAT = process.env.FUNCTION_DEFINITION_FORMAT || FunctionDefinitionFormat.ANTHROPIC;
   const ALLOWED_APPS_ONLY = options?.allowedAppsOnly ?? process.env.ACI_ALLOWED_APPS_ONLY === 'true';
@@ -27,8 +21,9 @@ export const createAciMCPConfig = (options?: {
     console.error('ACI_API_KEY not found in environment variables');
     throw new Error('ACI_API_KEY not found in environment variables');
   }
+  validateApiKey(process.env.ACI_API_KEY);
 
-  return {
+  config = {
     ACI_API_KEY: process.env.ACI_API_KEY,
     ACI_SERVER_URL,
     LINKED_ACCOUNT_OWNER_ID,
@@ -37,14 +32,9 @@ export const createAciMCPConfig = (options?: {
   };
 };
 
-let config: AciMCPConfig;
-
-export const initConfig = (options?: { allowedAppsOnly?: boolean; linkedAccountOwnerId?: string }): void => {
-  config = createAciMCPConfig(options);
-};
-
 export const getConfig = (): AciMCPConfig => {
   if (!config) {
+    console.error('Config not initialized. Call initConfig() first.');
     throw new Error('Config not initialized. Call initConfig() first.');
   }
   return config;
